Changed some multitransversality stuff
Move the remark about n-ary transversality not implying iterated transversality for maps of manifolds, and added a proposition showing that some transversality assumptions still imply others. I think this will be needed for working on showing that the the intersection map takes the geometric cup product to the cubical cup product.